Backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
chromadb
pypdf
tiktoken
requests
python-multipart
python-dotenv
